Update: add game logic to script.js
This adds 3 functions to the script.js files that facilitate playing
a game of rock, paper, scissors.

computerPlay: randomly generates a selection for the computer.

playRound: compares the computers selection against the players
selection and returns the results of the game.

game: plays a 5 round game, tallies the scores, displays those scores
at the end, and determines the winner of the game.