Update: makes many changes detailed below
Adds a playerPlay function that loops until the player makes
a valid selection.

Changes the return value of both computerPlay and playerPlay to
numeric values that represent the index of the selection in
POSSIBLE_PLAYS.

Simplifies how the winner of a round is determined and what
message about the round is returned with the results.

Makes WIN and LOSE results constant for less error prone comparisons.

Combines player and computer scores in game into a single score
state object.